Moved useSettersAsRefEventHandler to a separate file
