+++ src/index.ts
@@ -2,8 +2,9 @@ export { useForm } from './useForm';
 export { useFields } from './useFields';
 export { useErrors } from './useErrors';
 export {
-  useHandlers, useSettersAsEventHandler, useSettersAsRefEventHandler, useValidateAsSetter,
+  useHandlers, useSettersAsEventHandler, useValidateAsSetter,
 } from './useHandlers';
+export { useSettersAsRefEventHandler } from './useSettersAsRefEventHandler';
 export { useSubmission } from './useSubmission';
 export { useTouch } from './useTouch';
 export { useValidation } from './useValidation';
